backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pylint==3.0.2
pydantic==2.6.1
typing-extensions==4.9.0
requests==2.31.0
python-dotenv==1.0.1
groq 